Added strk_gas_price dir.
