+++ crates/blockifier/src/fee.rs
@@ -1,4 +1,3 @@
-pub mod errors;
 pub mod eth_gas_constants;
 pub mod fee_utils;
 pub mod gas_usage;
